rename file, adjust import
