+++ api-rework/invasives/api/viewsets/__init__.py
@@ -3,5 +3,5 @@ from .code import CodeViewSet
 from .ids_within_bounds import IdsWithinBoundsViewSet
 from .maps import MapGenerationRecordViewset, MapGenerationRequestViewSet
 from .migration import MigrationStatusViewSet
-from .recordset_rows import RecordsetRowsViewSet
+from .recordset import RecordsetRowsViewSet
 from .vector_tiles import VectorTileViewset
